修复DFS保存大文件时索引报错

+++ components/dfs/dfs_v1/src/dfs_posix.c
@@ -286,7 +286,7 @@ RTM_EXPORT(write);
  */
 off_t lseek(int fd, off_t offset, int whence)
 {
-    int result;
+    off_t result;
     struct dfs_file *d;
 
     d = fd_get(fd);
